Now using pyfomod to generate the xml files. Refactor class, added factory method. Keeping legacy executable generateFomod.py for integration purposes
