--- installer/templates/info.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0"?>
-<fomod>
-    <Name>$name</Name>
-
-    <Author>SlavicPotato(dev), KwekMaster(qa,FOMod)</Author>
-
-    <Version MachineVersion="$version">
-        $version
-    </Version>
-
-    <Description>
-       PotatoCBP is a advanced version of CBP Physics for Skyrim SE by polygonhell
-    </Description>
-
-    <Website>
-        https://github.com/SlavicPotato/CBPSSE
-    </Website>
-</fomod>
