--- installer/templates/plugin.xml
@@ -1,12 +1,0 @@
-                <plugin name="$pluginName">
-                    <description>
-                      $pluginDescription
-                    </description>
-                    <image path=""/>
-                    <files>
-                        <folder source="$pluginSource" destination="$pluginDestination"/>
-                    </files>
-                    <typeDescriptor>
-                        <type name="$isPluginRequired"/>
-                    </typeDescriptor>
-                </plugin>
